+++ packages/ag-charts-types/src/main.ts
@@ -1,11 +1,9 @@
-export * from './api/presetOptions';
 export * from './api/initialStateOptions';
 export * from './chart/annotationsOptions';
 export * from './chart/axisOptions';
 export * from './chart/callbackOptions';
 export * from './chart/crosshairOptions';
 export * from './chart/chartOptions';
-export * from './chart/chartBuilderOptions';
 export * from './chart/contextMenuOptions';
 export * from './chart/crossLineOptions';
 export * from './chart/dropShadowOptions';
@@ -23,6 +21,7 @@ export * from './chart/tooltipOptions';
 export * from './chart/themeOptions';
 export * from './chart/types';
 export * from './chart/zoomOptions';
+export * from './presets/presetOptions';
 export * from './series/seriesOptions';
 export * from './series/cartesian/cartesianOptions';
 export * from './series/cartesian/cartesianSeriesTooltipOptions';
@@ -65,8 +64,6 @@ export * from './series/topology/mapLineBackgroundOptions';
 export * from './series/flow-proportion/flowProportionOptions';
 export * from './series/flow-proportion/chordOptions';
 export * from './series/flow-proportion/sankeyOptions';
-export * from './series/gauge/gaugeCommonOptions';
-export * from './series/gauge/linearGaugeOptions';
-export * from './series/gauge/radialGaugeOptions';
 export * from './series/interpolationOptions';
 export * from './series/markerOptions';
+export * from './chartBuilderOptions';
