+++ packages/ag-charts-types/src/presets/gauge/gaugeCommonOptions.ts
